fix(frontend): hide mojang java 8 in macOS and 11 downloads (no target resource), refactor backend code style
