+++ src-tauri/src/launcher_config/helpers/java.rs
@@ -1,11 +1,16 @@
-use crate::error::SJMCLResult;
+use crate::error::{SJMCLError, SJMCLResult};
 use crate::launcher_config::models::{JavaInfo, LauncherConfig};
+use crate::resource::helpers::misc::{get_download_api, get_source_priority_list};
+use crate::resource::models::ResourceType;
+use crate::tasks::{download::DownloadParam, PTaskParam};
+use serde_json::Value;
 use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::sync::Mutex;
 use tauri::{AppHandle, Manager};
+use tauri_plugin_http::reqwest;
 
 #[cfg(target_os = "windows")]
 use std::error::Error;
@@ -429,4 +434,72 @@ pub fn parse_java_major_version(full_version: &str) -> (i32, bool) {
 
   let is_lts = [8, 11, 17, 21, 25].contains(&major_version);
   (major_version, is_lts)
+}
+
+pub async fn build_mojang_java_download_params(
+  app: &AppHandle,
+  version: &str,
+) -> SJMCLResult<Vec<PTaskParam>> {
+  let config = app.state::<Mutex<LauncherConfig>>().lock()?.clone();
+  let client = app.state::<reqwest::Client>();
+
+  let platform = match (&*config.basic_info.os_type, &*config.basic_info.arch) {
+    ("windows", "aarch64") => "windows-arm64",
+    ("windows", "x86_64") => "windows-x64",
+    ("windows", _) => "windows-x86",
+    ("macos", "aarch64") => "mac-os-arm64",
+    ("macos", _) => "mac-os",
+    ("linux", "x86_64") => "linux",
+    _ => "linux-i386",
+  };
+  let runtime_type = match version {
+    "8" => "jre-legacy",
+    "21" => "java-runtime-delta",
+    _ => "java-runtime-gamma",
+  };
+
+  let priority_list = get_source_priority_list(&config);
+  let mut json: Option<Value> = None;
+
+  for source_type in priority_list.iter() {
+    if let Ok(api_url) = get_download_api(*source_type, ResourceType::MojangJava) {
+      if let Ok(response) = client.get(api_url).send().await {
+        if let Ok(parsed_json) = response.json::<Value>().await {
+          json = Some(parsed_json);
+          break;
+        }
+      }
+    }
+  }
+
+  let json =
+    json.ok_or_else(|| SJMCLError("Failed to fetch Mojang Java runtime manifest".into()))?;
+  let manifest_url = json[platform][runtime_type][0]["manifest"]["url"]
+    .as_str()
+    .ok_or_else(|| SJMCLError("Failed to parse manifest URL".into()))?;
+
+  let manifest: Value = client.get(manifest_url).send().await?.json().await?;
+  let runtime_dir = app.path().resolve(
+    format!("runtime/java-{}", version),
+    tauri::path::BaseDirectory::AppData,
+  )?;
+
+  let download_params: Vec<_> = manifest["files"]
+    .as_object()
+    .ok_or_else(|| SJMCLError("Invalid files data".into()))?
+    .iter()
+    .filter_map(|(path, info)| {
+      let raw = info["downloads"]["raw"].as_object()?;
+      let (url, sha1) = (raw["url"].as_str()?, raw["sha1"].as_str()?);
+
+      Some(PTaskParam::Download(DownloadParam {
+        src: url.parse().ok()?,
+        dest: runtime_dir.join(path),
+        filename: None,
+        sha1: Some(sha1.into()),
+      }))
+    })
+    .collect();
+
+  Ok(download_params)
 }
